backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2